Add: マスの種類によってプレイヤーに送るリクエストをハンドルするようにしました
